remove a dozen-plus identical copies of pixgauss2d
